Struct identity_iota::client::Client

pub struct Client {
    pub(crate) client: Client,
    pub(crate) network: Network,
}

Fields
client: Clientnetwork: Network
Implementations

impl Client

pub fn new() -> Result<Self>

Creates a new Client with default settings.

pub fn builder() -> ClientBuilder

Creates a ClientBuilder to configure a new Client.

This is the same as ClientBuilder::new().

pub fn from_network(network: Network) -> Result<Self>

Creates a new Client with default settings for the given Network.

pub fn from_builder(builder: ClientBuilder) -> Result<Self>

Creates a new Client based on the ClientBuilder configuration.

pub fn network(&self) -> Network

Returns the Client Tangle network.

pub fn default_node_url(&self) -> &'static Url

Returns the default node URL of the Client network.

pub fn explorer_url(&self) -> &'static Url

Returns the web explorer URL of the Client network.

pub fn transaction_url(&self, transaction: &BundledTransaction) -> Url

Returns the web explorer URL of the given transaction.

pub fn transaction_hash(&self, transaction: &BundledTransaction) -> String

Returns the hash of the Tangle transaction as a tryte-encoded String.

pub async fn publish_document<'_, '_>(
&'_ self,
document: &'_ IotaDocument
) -> Result<BundledTransaction>

Publishes an IotaDocument to the Tangle.

Note: The only validation performed is to ensure the correct Tangle network is selected.

pub async fn publish_diff<'_, '_, '_>(
&'_ self,
message_id: &'_ MessageId,
diff: &'_ DocumentDiff
) -> Result<BundledTransaction>

Publishes a DocumentDiff to the Tangle.

Note: The only validation performed is to ensure the correct Tangle network is selected.

pub async fn read_document<'_, '_>(
&'_ self,
did: &'_ IotaDID
) -> Result<IotaDocument>

pub async fn read_document_chain<'_, '_>(
&'_ self,
did: &'_ IotaDID
) -> Result<DocumentChain>

pub async fn read_messages<'_, '_>(
&'_ self,
address: &'_ str
) -> Result<Vec<Message>>

pub async fn send_transfer<'_>(
&'_ self,
transfer: Transfer
) -> Result<BundledTransaction>

pub fn check_network(&self, did: &IotaDID) -> Result<()>