Struct identity_iota::credential::validator::CredentialValidator

pub struct CredentialValidator<'a> {
    client: &'a Client,
}

Fields
client: &'a Client
Implementations

impl<'a> CredentialValidator<'a>

pub const fn new(client: &'a Client) -> Self

Creates a new CredentialValidator.

pub async fn check<T, '_, '_>(
&'_ self,
data: &'_ str
) -> Result<CredentialValidation<T>> where
T: DeserializeOwned + Serialize,

Deserializes the given JSON-encoded VerifiableCredential and validates all associated DID documents.

pub async fn check_presentation<T, U, '_, '_>(
&'_ self,
data: &'_ str
) -> Result<PresentationValidation<T, U>> where
T: Clone + DeserializeOwned + Serialize,
U: Clone + DeserializeOwned + Serialize,

Deserializes the given JSON-encoded VerifiablePresentation and validates all associated DID documents/VerifiableCredentials.

pub async fn validate_credential<T, '_>(
&'_ self,
credential: VerifiableCredential<T>
) -> Result<CredentialValidation<T>> where
T: Serialize,

Validates the VerifiableCredential proof and all relevant DID documents.

Note: The credential is expected to have a proof created by the issuing party. Note: The credential issuer URL is expected to be a valid DID. Note: Credential subject IDs are expected to be valid DIDs (if present).

pub async fn validate_presentation<T, U, '_>(
&'_ self,
presentation: VerifiablePresentation<T, U>
) -> Result<PresentationValidation<T, U>> where
T: Clone + Serialize,
U: Clone + Serialize,

Validates the VerifiablePresentation proof and all relevant DID documents.

Note: The presentation holder is expected to be a valid DID. Note: The presentation is expected to have a proof created by the holder.

async fn validate_document<'_, '_>(
&'_ self,
did: &'_ str
) -> Result<DocumentValidation>