Struct identity_iota::did::Document

pub struct Document {
    document: CoreDocument<VerifiableProperties<BaseProperties>, Object, ()>,
    message_id: MessageId,
}

A DID Document adhering to the IOTA DID method specification.

This is a thin wrapper around the Document type from the identity_did crate.

Fields
document: CoreDocument<VerifiableProperties<BaseProperties>, Object, ()>message_id: MessageId
Implementations

impl Document

pub fn from_keypair(keypair: &KeyPair) -> Result<Self>

Creates a new DID Document from the given KeyPair.

The DID Document will be pre-populated with a single authentication method based on the provided KeyPair.

The authentication method will have the DID URL fragment #authentication and can be easily retrieved with Document::authentication.

pub fn from_authentication(method: Method) -> Result<Self>

Creates a new DID Document from the given verification method.

pub unsafe fn from_authentication_unchecked(method: Method) -> Self

Creates a new DID Document from the given verification method without performing validation checks.

Safety

This must be guaranteed safe by the caller.

pub fn try_from_core(document: CoreDocument) -> Result<Self>

Converts a generic DID Document to an IOTA DID Document.

Errors

Returns Err if the document is not a valid IOTA DID Document.

fn check_authentication(method: &CoreMethod) -> Result<()>

pub fn as_document(
&self
) -> &CoreDocument<VerifiableProperties<BaseProperties>, Object, ()>

Returns a reference to the underlying Document.

pub unsafe fn as_document_mut(
&mut self
) -> &mut CoreDocument<VerifiableProperties<BaseProperties>, Object, ()>

Returns a mutable reference to the underlying Document.

Safety

This function is unsafe because it does not check that modifications made to the Document maintain a valid IOTA DID Document.

If this constraint is violated, it may cause issues with future uses of the DID Document.

pub fn id(&self) -> &DID

Returns the DID document id.

pub fn authentication(&self) -> &Method

Returns the default authentication method of the DID document.

fn authentication_id(&self) -> &str

pub fn created(&self) -> Timestamp

Returns the timestamp of when the DID document was created.

pub fn set_created(&mut self, value: Timestamp)

Sets the timestamp of when the DID document was created.

pub fn updated(&self) -> Timestamp

Returns the timestamp of the last DID document update.

pub fn set_updated(&mut self, value: Timestamp)

Sets the timestamp of the last DID document update.

pub fn previous_message_id(&self) -> &MessageId

Returns the Tangle message id of the previous DID document, if any.

pub fn set_previous_message_id(&mut self, value: impl Into<MessageId>)

Sets the Tangle message id the previous DID document.

pub fn immutable(&self) -> bool

Returns true if the DID Document is flagged as immutable.

pub fn set_immutable(&mut self, value: bool)

Sets the value of the immutable flag.

pub fn properties(&self) -> &Object

Returns a reference to the custom DID Document properties.

pub fn properties_mut(&mut self) -> &mut Object

Returns a mutable reference to the custom DID Document properties.

pub fn insert_method(&mut self, scope: MethodScope, method: Method) -> bool

Adds a new Verification Method to the DID Document.

pub fn remove_method(&mut self, did: &DID) -> Result<()>

Removes all references to the specified Verification Method.

pub fn sign(&mut self, secret: &SecretKey) -> Result<()>

Signs the DID document with the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the signature operation fails.

pub fn verify(&self) -> Result<()>

Verifies the signature of the DID document.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the verification operation fails.

pub fn sign_data<X>(&self, data: &mut X, secret: &SecretKey) -> Result<()> where
X: Serialize + SetSignature,

Signs the provided data with the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the signature operation fails.

pub fn verify_data<X>(&self, data: &X) -> Result<()> where
X: Serialize + TrySignature,

Verifies the signature of the provided data.

Note: It is assumed that the signature was created using a verification method contained within the DID Document.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the verification operation fails.

pub fn diff(
&self,
other: &Self,
message: MessageId,
secret: &SecretKey
) -> Result<DocumentDiff>

Creates a DocumentDiff representing the changes between self and other.

The returned DocumentDiff will have a digital signature created using the default authentication method and secret.

Errors

Fails if the diff operation or signature operation fails.

pub fn merge(&mut self, diff: &DocumentDiff) -> Result<()>

Verifies a DocumentDiff signature and merges the changes into self.

If merging fails self remains unmodified, otherwise self represents the merged document state.

Errors

Fails if the merge operation or signature operation fails.

pub async fn publish<'client, C, '_>(&'_ mut self, client: C) -> Result<()> where
C: Into<Option<&'client Client>>,

Publishes the DID Document to the Tangle

Uses the provided client or a default Client based on the DID network.

pub fn diff_address(message_id: &MessageId) -> Result<String>

Returns the Tangle address of the DID diff chain.