Struct identity_iota::did::IotaDocument

pub struct IotaDocument {
    document: Document<VerifiableProperties<BaseProperties>, (), ()>,
    message_id: MessageId,
}

Fields
document: Document<VerifiableProperties<BaseProperties>, (), ()>message_id: MessageId
Implementations

impl IotaDocument

pub fn generate_ed25519<'a, 'b, T, U>(
tag: &str,
network: T,
shard: U
) -> Result<(Self, KeyPair)> where
T: Into<Option<&'a str>>,
U: Into<Option<&'b str>>,

pub fn try_from_document(document: Document<Object>) -> Result<Self>

Converts a generic DID Document to an IotaDocument.

Errors

Returns Err if the document is not a valid IotaDocument.

pub fn builder() -> IotaDocumentBuilder

Creates a IotaDocumentBuilder to configure a new IotaDocument.

This is the same as IotaDocumentBuilder::new().

pub fn id(&self) -> &IotaDID

Returns the DID document id.

pub fn authentication(&self) -> MethodWrap<'_>

Returns the default authentication method of the DID document.

pub fn authentication_bytes(&self) -> Result<Vec<u8>>

Returns the key bytes of the default DID document authentication method.

pub fn authentication_type(&self) -> MethodType

Returns the method type of the default DID document authentication method.

pub fn created(&self) -> Timestamp

Returns the timestamp of when the DID document was created.

pub fn set_created(&mut self, value: Timestamp)

Sets the timestamp of when the DID document was created.

pub fn set_created_now(&mut self)

Sets the DID document "created" timestamp to Timestamp::now.

pub fn updated(&self) -> Timestamp

Returns the timestamp of the last DID document update.

pub fn set_updated(&mut self, value: Timestamp)

Sets the timestamp of the last DID document update.

pub fn set_updated_now(&mut self)

Sets the DID document "updated" timestamp to Timestamp::now.

pub fn previous_message_id(&self) -> &MessageId

Returns the Tangle message id of the previous DID document, if any.

pub fn set_previous_message_id<T>(&mut self, value: T) where
T: Into<MessageId>,

Sets the Tangle message id the previous DID document.

pub fn immutable(&self) -> bool

Returns true if the IotaDocument is flagged as immutable.

pub fn set_immutable(&mut self, value: bool)

Sets the value of the immutable flag.

pub fn properties(&self) -> &Object

Returns a reference to the custom IotaDocument properties.

pub fn properties_mut(&mut self) -> &mut Object

Returns a mutable reference to the custom IotaDocument properties.

pub fn as_document(
&self
) -> &Document<VerifiableProperties<BaseProperties>, (), ()>

Returns a reference to the underlying Document.

pub unsafe fn as_document_mut(
&mut self
) -> &mut Document<VerifiableProperties<BaseProperties>, (), ()>

Returns a mutable reference to the underlying Document.

Safety

This function is unsafe because it does not check that modifications made to the Document maintain a valid IotaDocument.

If this constraint is violated, it may cause issues with future uses of the IotaDocument.

pub fn sign(&mut self, secret: &SecretKey) -> Result<()>

Signs the DID document with the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the signature operation fails.

pub fn verify(&self) -> Result<()>

Verifies the signature of the DID document.

Note: It is assumed that the signature was created using the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the verification operation fails.

pub fn sign_data<T>(&self, data: &mut T, secret: &SecretKey) -> Result<()> where
T: Serialize + SetSignature,

Signs the provided data with the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the signature operation fails.

pub fn verify_data<T>(&self, data: &T) -> Result<()> where
T: Serialize + TrySignature,

Verfies the signature of the provided data.

Note: It is assumed that the signature was created using the default authentication method.

Errors

Fails if an unsupported verification method is used, document serialization fails, or the verification operation fails.

pub fn resolve_options(&self) -> Result<SignatureOptions>

pub fn diff(
&self,
other: &Self,
secret: &SecretKey,
previous_message_id: MessageId
) -> Result<DocumentDiff>

Creates a DocumentDiff representing the changes between self and other.

The returned DocumentDiff will have a digital signature created using the default authentication method and secret.

Errors

Fails if the diff operation or signature operation fails.

pub fn merge(&mut self, diff: &DocumentDiff) -> Result<()>

Verifies a DocumentDiff signature and merges the changes into self.

If merging fails self remains unmodified, otherwise self represents the merged document state.

Errors

Fails if the merge operation or signature operation fails.

pub async fn publish<'_>(&'_ mut self) -> Result<()>

Publishes the IotaDocument to the Tangle using a default Client.

pub async fn publish_with_client<'_, '_>(
&'_ mut self,
client: &'_ Client
) -> Result<()>

Publishes the IotaDocument to the Tangle using the provided Client.

pub fn diff_address(message_id: &MessageId) -> Result<String>

Returns the Tangle address of the DID diff chain.